worker: implement MR remarks
